Backend/requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.4.0
django-environ==0.11.2
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gcloud==0.17.0
googleapis-common-protos==1.63.2
httplib2==0.22.0
idna==3.7
jws==0.1.3
oauth2client==3.0.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycryptodome==3.4.3
PyJWT==2.9.0
pyparsing==3.1.2
Pyrebase==3.0.27
python-dotenv==1.0.1
python-jwt==2.0.1
redis==5.0.6
requests==2.32.3
requests-toolbelt==0.7.0
rsa==4.9
six==1.16.0
spotipy==2.24.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
